Repo creation rule now properly depends on the wheel library
This requires the repo creation tool to be a bundled executable instead of just
a script, and since it's used for a repository rule, that means the bundled
executable needs to be source controlled.

As a result, pretty much re-organized and re-named everything.

Still need to create a rule or script to update the repo creation tool.