Ditched the pep425 tags
It added too much complexity and didn't really work in practice because it
resulted in a single named repository having potentially different external
dependencies depending on the Python runtime.

Instead, requirements files are built or selected based on Bazel platform, and
a different repository name should be used for Python 2/3 (e.g. 'pip2' and 'pip3')