During resolution, get details from dist before moving built wheel
Some distributions, like salt, have a lot of logic in setup.py that may cause a
"universal" wheel built with Python 2 to actually have different details, like
requirements, than a "universal" wheel built with Python 3. This seems
inherently broken on its own, but it means that the same universal wheel can't
be used to determine the list of dependencies for all environments. Instead,
during resolution, the dist information should be taken from the temporary local
wheel before performing the move-if-necessary to the destination wheel
directory. That way, the dist information will correctly pertain to the
environment in which it was built.