The pip_lock rule no longer builds universal wheels
A given distribution can produce a slightly different wheel in each environment, despite
the fact that it was built using the universal flag, due to the fact that setup.py can
execute arbitrary code. As such, any distribution for which the pip_lock rule builds
wheels should not be built universally, just in case. Extra disk space. Oh well.